seperate Coinbalance component and amount component
